moved asm and user to their own file
