+++ projects/schematics/src/shared/libs-constants.ts
@@ -1,5 +1,3 @@
-import { FeatureConfig } from './utils/lib-utils';
-
 export const PRODUCT_CONFIGURATOR_RULEBASED_FEATURE =
   'productConfiguratorRulebased';
 export const PRODUCT_CONFIGURATOR_TEXTFIELD_FEATURE =
@@ -169,38 +167,6 @@ export const CLI_EPD_VISUALIZATION_FEATURE = 'EPD-Visualization';
 /***** CLI end *****/
 
 /***** Feature libs configuration start *****/
-export const ASM_FOLDER_NAME = 'asm';
-// TODO:#schematics - rename to `ASM_FEATURE_MODULE_NAME`
-export const ASM_MODULE_NAME = 'Asm';
-
-export const ASM_FEATURE_NAME_CONSTANT = 'ASM_FEATURE';
-export const ASM_MODULE = 'AsmModule';
-export const ASM_ROOT_MODULE = 'AsmRootModule';
-export const ASM_TRANSLATIONS = 'asmTranslations';
-export const ASM_TRANSLATION_CHUNKS_CONFIG = 'asmTranslationChunksConfig';
-
-export const ASM_SCHEMATICS_CONFIG: FeatureConfig = {
-  folderName: ASM_FOLDER_NAME,
-  moduleName: ASM_MODULE_NAME,
-  featureModule: {
-    name: ASM_MODULE,
-    importPath: SPARTACUS_ASM,
-  },
-  rootModule: {
-    name: ASM_ROOT_MODULE,
-    importPath: SPARTACUS_ASM_ROOT,
-  },
-  lazyLoadingChunk: {
-    moduleSpecifier: SPARTACUS_ASM_ROOT,
-    namedImports: [ASM_FEATURE_NAME_CONSTANT],
-  },
-  i18n: {
-    resources: ASM_TRANSLATIONS,
-    chunks: ASM_TRANSLATION_CHUNKS_CONFIG,
-    importPath: SPARTACUS_ASM_ASSETS,
-  },
-};
-
 export const CART_BASE_MODULE = 'CartBaseModule';
 export const CART_BASE_ROOT_MODULE = 'CartBaseRootModule';
 export const MINI_CART_MODULE = 'MiniCartModule';
@@ -268,81 +234,6 @@ export const TMS_GTM_MODULE = 'GtmModule';
 export const TMS_AEP_MODULE = 'AepModule';
 export const PERSONALIZATION_MODULE = 'PersonalizationModule';
 export const PERSONALIZATION_ROOT_MODULE = 'PersonalizationRootModule';
-
-export const USER_FOLDER_NAME = 'user';
-export const SCSS_FILE_NAME = 'user.scss';
-// TODO:#schematics - rename to `ASM_FEATURE_MODULE_NAME`
-export const USER_MODULE_NAME = 'User';
-
-export const USER_ACCOUNT_FEATURE_NAME_CONSTANT = 'USER_ACCOUNT_FEATURE';
-export const USER_ACCOUNT_MODULE = 'UserAccountModule';
-export const USER_ACCOUNT_ROOT_MODULE = 'UserAccountRootModule';
-export const USER_ACCOUNT_TRANSLATIONS = 'userAccountTranslations';
-export const USER_ACCOUNT_TRANSLATION_CHUNKS_CONFIG =
-  'userAccountTranslationChunksConfig';
-export const USER_ACCOUNT_SCHEMATICS_CONFIG: FeatureConfig = {
-  folderName: USER_FOLDER_NAME,
-  moduleName: USER_MODULE_NAME,
-  featureModule: {
-    name: USER_ACCOUNT_MODULE,
-    importPath: SPARTACUS_USER_ACCOUNT,
-  },
-  rootModule: {
-    name: USER_ACCOUNT_ROOT_MODULE,
-    importPath: SPARTACUS_USER_ACCOUNT_ROOT,
-  },
-  lazyLoadingChunk: {
-    moduleSpecifier: SPARTACUS_USER_ACCOUNT_ROOT,
-    namedImports: [USER_ACCOUNT_FEATURE_NAME_CONSTANT],
-  },
-  i18n: {
-    resources: USER_ACCOUNT_TRANSLATIONS,
-    chunks: USER_ACCOUNT_TRANSLATION_CHUNKS_CONFIG,
-    importPath: SPARTACUS_USER_ACCOUNT_ASSETS,
-  },
-  styles: {
-    scssFileName: SCSS_FILE_NAME,
-    importStyle: SPARTACUS_USER,
-  },
-};
-
-export const USER_PROFILE_FEATURE_NAME_CONSTANT = 'USER_PROFILE_FEATURE';
-export const USER_PROFILE_MODULE = 'UserProfileModule';
-export const USER_PROFILE_ROOT_MODULE = 'UserProfileRootModule';
-export const USER_PROFILE_TRANSLATIONS = 'userProfileTranslations';
-export const USER_PROFILE_TRANSLATION_CHUNKS_CONFIG =
-  'userProfileTranslationChunksConfig';
-export const USER_PROFILE_SCHEMATICS_CONFIG: FeatureConfig = {
-  folderName: USER_FOLDER_NAME,
-  moduleName: USER_MODULE_NAME,
-  featureModule: {
-    name: USER_PROFILE_MODULE,
-    importPath: SPARTACUS_USER_PROFILE,
-  },
-  rootModule: {
-    name: USER_PROFILE_ROOT_MODULE,
-    importPath: SPARTACUS_USER_PROFILE_ROOT,
-  },
-  lazyLoadingChunk: {
-    moduleSpecifier: SPARTACUS_USER_PROFILE_ROOT,
-    namedImports: [USER_PROFILE_FEATURE_NAME_CONSTANT],
-  },
-  i18n: {
-    resources: USER_PROFILE_TRANSLATIONS,
-    chunks: USER_PROFILE_TRANSLATION_CHUNKS_CONFIG,
-    importPath: SPARTACUS_USER_PROFILE_ASSETS,
-  },
-  styles: {
-    scssFileName: SCSS_FILE_NAME,
-    importStyle: SPARTACUS_USER,
-  },
-  dependencyManagement: {
-    featureName: CLI_USER_PROFILE_FEATURE,
-    featureDependencies: {
-      [SPARTACUS_USER]: [CLI_USER_ACCOUNT_FEATURE],
-    },
-  },
-};
 
 export const CDC_MODULE = 'CdcModule';
 export const CDC_ROOT_MODULE = 'CdcRootModule';
