+++ projects/schematics/src/shared/updateable-constants.ts
@@ -1,10 +1,18 @@
 import {
+  ASM_MODULE,
+  ASM_ROOT_MODULE,
+} from './lib-configs/asm-schematics-config';
+import {
+  USER_ACCOUNT_MODULE,
+  USER_ACCOUNT_ROOT_MODULE,
+  USER_PROFILE_MODULE,
+  USER_PROFILE_ROOT_MODULE,
+} from './lib-configs/user-schematics-config';
+import {
   ADD_TO_CART_MODULE,
   ADD_TO_WISHLIST_MODULE,
   ADMINISTRATION_MODULE,
   ADMINISTRATION_ROOT_MODULE,
-  ASM_MODULE,
-  ASM_ROOT_MODULE,
   BULK_PRICING_MODULE,
   BULK_PRICING_ROOT_MODULE,
   CART_BASE_MODULE,
@@ -99,10 +107,6 @@ import {
   TMS_AEP_MODULE,
   TMS_BASE_MODULE,
   TMS_GTM_MODULE,
-  USER_ACCOUNT_MODULE,
-  USER_ACCOUNT_ROOT_MODULE,
-  USER_PROFILE_MODULE,
-  USER_PROFILE_ROOT_MODULE,
   VARIANTS_MODULE,
   VARIANTS_ROOT_MODULE,
 } from './libs-constants';
